requirements.txt

Flask==1.0.2
Flask-Cors==3.0.2
gunicorn==19.7.1
numpy~=1.23.4
pickleshare==0.7.4
scipy==1.0.0



PySide6~=6.4.0.1
Pillow~=9.3.0